move new constituency data converter to a class
